// orval/src/model/launchpadCollection.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Indexer API
 * The Querys for the NFT Indexer
 * OpenAPI spec version: 1.0
 */
import type { WhiteListAddress } from './whiteListAddress';

export interface LaunchpadCollection {
  launchId: string;
  collectionType: string;
  native: boolean;
  kioskStandard: boolean;
  name: string;
  description: string;
  supply: string;
  price: number;
  maxItemPerAddress: string;
  startTime: string;
  whiteListEnabled: boolean;
  whiteListStartTime: string;
  whiteListPrice: number;
  whiteListSupply: string;
  whiteListAddresses: WhiteListAddress[];
  customEnabled: boolean;
  customName: string;
  customStartTime: string;
  customPrice: number;
  customSupply: string;
  mintCount: number;
  state: string;
  createdAt: string;
  updatedAt: string;
}